Rename ModbusSniffer to ModbusSpy
